media: Reference count and power handling
Basically these are the interface functions:

media_entity_get() - acquire entity
media_entity_put() - release entity

If the entity is of node type, the power change is distributed to all
connected entities. For non-nodes it only affects that very node. A
mutex is used to serialise access to the entity graph.

In the background there's a depth-first search algorithm that traverses
the enabled links in the graph. All these functions parse the graph to
implement whatever they're to do.

The module counters are increased/decreased in media_entity_get/put to
prevent module unloading when an entity is referenced.

Signed-off-by: Sakari Ailus <sakari.ailus@maxwell.research.nokia.com>
Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Stanimir Varbanov <svarbanov@mm-sol.com>