rx51 cam: Drop power control from board file
This patch is in preparation to move to the regulator framework. We
should no more access the TWL5030 registers directly, so drop all power
handling from board files.

The handling of the VANA signal for the camera should be done by the
camera sensor and lens driver. And the handling of the VDDA_VCSI signal
should be done by the ISP driver.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>